@@ -22,11 +22,12 @@ classifiers =
     #Development Status :: 5 - Production/Stable
     Programming Language :: Python :: 3 :: Only
     Programming Language :: Python :: 3
-    Programming Language :: Python :: 3.5
     Programming Language :: Python :: 3.6
     Programming Language :: Python :: 3.7
     Programming Language :: Python :: 3.8
     Programming Language :: Python :: 3.9
+    Programming Language :: Python :: Implementation :: CPython
+    Programming Language :: Python :: Implementation :: PyPy
     License :: OSI Approved :: MIT License
     Environment :: Console
     Intended Audience :: Developers
@@ -42,11 +43,11 @@ project_urls =
 packages = find:
 package_dir =
     =src
-python_requires = ~=3.5
+python_requires = ~=3.6
 install_requires =
-    click~=7.0
-    packaging>=16
-    requests==2.*
+    click     ~= 7.0
+    packaging >= 16
+    requests  ~= 2.20
 
 [options.packages.find]
 where = src
